Adding a maxJdkVersion (1.x form) option to ease the configuration.
This should ease for example sharing a property that could be used both
for the target jdk in the maven-compiler-plugin and this enforcer rule.

This is still possible to use maxJavaMajorVersionNumber for more advanced
use cases if needed. I guess most people just won't use it.

Added also the corresponding IT.