MVERSIONS-249: implemented updatePropertyVersions.
If the 'updatePropertyVersions' property is defined, the properties defining a version matching the
dependencies of the remote POM are considered. If a particular property defines the version of several
artifacts, all artifacts should be defined by the remote POM and have one and only one version. In other
cases, the property is not updated.

In reportMode, the differences are logged alongside the usual report. If reportMode is disabled, the
pom is updated with the differences.